userpage pull to refresh
