+++ lib/pages/my_page/my_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:project_finaltask/components/appbar_component.dart';
-import 'package:project_finaltask/views/user_view.dart';
+import 'package:project_finaltask/views/user_view/user_view.dart';
 
 import '../../models/user.dart';
 import '../../qiita_client.dart';
